refactor:apis에 로직 분리 => 사용하는 곳에서 처리하게 이동 및 apis에는 정의만 하게 수정 & 지역코드 common폴더의 region을 사용하도록 수정
